Improved the IMovable interface.
Classes implementing IMovable now only have to implement two (important)
methods. getSpeedVector() needs to return the speed vector of the object,
which the PlayingField will use in order to automatically move that
object.
PlayingField now automatically moves the object using getSpeedVector() and
before moving, it calls the preMove() method in case the object wants to
adjust something (for example useful for AI).
Adjusted EnemyFish so that it doesn't throw compile errors with this
edited interface.