Hotfix for maven build for master.
The GUI tests keep the logger loaded, which meant that each time the a new
GUI test was ran, new handlers were added to Logger. This slows down the
maven build and caused the logger tests to fail, which where ran after the
GUI test in maven build. Cleaned up initiliazation of loggers + made sure
the TestLog class always uses a clean logger for it's tests.

<bug