Updated the PowerUp class for its future children
Splitsed the abstract onCollide method into a non-abstract onCollide
method and an abstract executeEffect(PlayerFish pfish). The executeEffect
method is automatically called when the other:ICollidable in the onCollide
method is a PlayerFish.

Added the field pfield : PlayingField to PowerUp, which can be accesed
using getPField(). This is useful for the child classes of PowerUp when
they implement the executeEffect method.

<organisation