Refactored SinglePlayerPlayingField
Instead of giving a SinglePlayerController to the
SinglePlayerPlayingField and having code mixed between those 2 classes,
the PlayingField now offers a playerProperty, which listeners can be added
to by any GUI component that needs it, removing the need for such
interactions.

Long story short: less interaction between classes, cleaner and clearer
code that is easier to understand.

<organisation,feature,interface,performance