Removed anything to do with class PuSmallFishMagnet
Changing the AI of the EnemyFish from outside that class made
the code too messy. Actually implementing it would require changes
in the code organisation (I have some ideas on how to do that),
but I'm unsure if I'm able to apply those changes. So therefore,
I decided to remove the attribute selfControlling in the EnemyFish
and completely remove the PuSmallFishMagnet class. Maybe later.

<bug, organisation