Refactored LevelBuilder to name EnemyFishFactory.
This refactoring is something that should have happened when we moved the
responsibilities of the LevelBuilder to other classes and were only left
with a LevelBuilder which had the only task of prescribing how EnemyFish
work in the game. Therefore this rename, as the class belongs to a Factory
Design pattern.

<organisation