Entities now have a behaviour as attribute.
Handling how an entity moves is the responsibility of the behaviour.
Everything else is the entity's problem now. hitWall and canMoveThrough
wall have more to do with entities than with behaviours, so instead of
having those methods in the IBehaviour interface, it is an abstract
method in Entity.
Some tests also had to be removed due to the big change, but they will
still need to be added once more. RandomBehaviour and FrozenBehaviour
are now two new types of behaviours.

TODO: letting the PlayerFish also have a Behaviour as an attribute,
instead of implementing it, which is ugly.