Refactor tests based on boundingbox refactor.
Changed two tests. Deleted the TestEnemyFish class (for now). Even though
the EnemyFish class did not use BoundingBoxes anymore for quite some time,
it's test class did still use them. With the removal of the intersects
method in BoundingBox these tests began to fail. Still to do: write new
tests for EnemyFish, based on the fact that EnemyFish has a CollisionMask
and not a BoundingBox.

<organisation